+++ server/src/test/java/game/world/WorldChunkTest.java
@@ -12,7 +12,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
 public class WorldChunkTest {
-    World world = new World();
+    final World world = new World();
 
     @Test
     void testNumberOfChunks() {
@@ -65,4 +65,16 @@ public class WorldChunkTest {
                 .orElseThrow();
     }
 
+    @Test
+    void testFindIntersectingChunks() {
+        final var worldChunkSize = world.getConfig().chunks.size;
+        final var originWorldChunk = world.chunks.findChunk(Vector.ORIGIN);
+        final var position = originWorldChunk.box.getCenter();
+        final double epsilon = 0.01;
+
+        var worldChunksInRadius = world.chunks.findIntersectingChunks(position, worldChunkSize / 2 - epsilon);
+        assertEquals(1, worldChunksInRadius.size());
+        worldChunksInRadius = world.chunks.findIntersectingChunks(position, worldChunkSize / 2 + epsilon);
+        assertEquals(5, worldChunksInRadius.size());
+    }
 }
